Split helper into multiple files
Not going to do the whole split-with-branch-for-each-file-thingy. Too lazy. Hope history won't be so important
