+++ src/helper/getLinkListPusher.ts
@@ -1,0 +1,6 @@
+export const getLinkListPusher = <T>(linkList: Array<T>): ((link: T) => T) =>
+  function (link) {
+    linkList.push(link)
+
+    return link
+  }
