+++ src/helper/getRandomInt.ts
@@ -1,0 +1,1 @@
+export const getRandomInt = (i: number): number => Math.floor(Math.random() * i)
