+++ src/scripts/comic/tablecomic-2.js
@@ -1,4 +1,4 @@
-import { multiplyColor } from '@/helper/helper'
+import { multiplyColor } from '@/helper/helperColor'
 
 // BEGINN getTableComic /\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-/\-
 export const getTableComic = function getTableComic(args) {
