+++ src/scripts/persons-and-trees/person-head.js
@@ -1,4 +1,4 @@
-import { mult, sub } from '@/helper/helper'
+import { mult, sub } from '@/helper/helperDim'
 
 import { Object } from './object.js'
